requirements/base.txt
flask-apispec==0.11.0
flask==2.0.3
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
flask-restful==0.3.9
pydantic==1.9.0
python-dotenv==0.20.0
requests==2.27.1